Major style updates and corrections
Also known as the You-Have-No-Sense-Of-Style update. It has been long
established that this is the case for me, in code, clothing, and so forth.
This commit contains a lot of major and minor style improvements, and some
general reorganization. Now that much of the initial hackery is done, it was
high time to make everything easier to read, debug, and use.

Significant changes in this commit are:
 * Lines now almost always wrap at 80 chars
 * Header info is formatted a little more nicely
 * Python scripts changed from concatenation to % string replacement
 * Database objects are handled a little less clumsily
 * Greatly improved logging, and added a "verbose" level
 * facade-worker.py is broken up into functions

The last item should make it easier to work on facade-worker.py, because you can
comment out stuff you aren't working on.  If you're modifying something in the
gitdm section, there's no reason to keep trying to update git repos every few
minutes.

There are a few other minor tweaks that should make things easier to use.

Signed-off-by: Brian Warner <brian@bdwarner.com>