Improved logging when fixing changed affiliations
Previously, affiliation corrections where somewhat of a black box.  Also, they
kinda didn't work right.  Both of these have been fixed.  Logging is now better
and it tells you when it's detected a change to the gitdm config files.  Also,
now it actually makes the change when a person worked for multiple companies,
which it should have been doing all along but wasn't due to a boneheaded
indentation error.

Signed-off-by: Brian Warner <brian@bdwarner.com>