Now importing repos from Gerrit in bulk
Seriously, Gerrit is a pain to scrape for repo links.  Every one is on a
different page.  So now we're going to try to be a little smarter about it.

But not too smart... authentication isn't supported and if the git repos have
different base URLs than the Gerrit instance, you're going to have to put those
in manually.

Signed-off-by: Brian Warner <brian@bdwarner.com>