Add ability to select metric (LoC added, LoC removed, patches, emails)
gitdm calculates a bunch of cool metrics, but up until now we were just using
LoC added. Of course you could export the CSV and calculate the others in Excel
using pivot tables, but that's, like, real work and a pain in the butt.

Now there's a selector at the top of each page to choose which metric should be
displayed in the summary tables.

Signed-off-by: Brian Warner <brian@bdwarner.com>