Added basic user authentication support
Added basic user authentication, to enable future support for administrators
vs. viewers. This will make it possible to put Facade somewhere that others
can view the statistics, but not add/remove projects and repos, or change the
global settings.

This patchset includes support for creating an administrator at setup (this
means that you need to run setup again), changing your email address, and
changing your password.  Passwords are hashed, salted, and compared using the
built-in functions in PHP.

Signed-off-by: Brian Warner <brian@bdwarner.com>