Fixed authentication issue.
I should not work on this while jetlagged. I should not work on this while
jetlagged. I should not work on this while jetlagged.

The real problem is that there was no session being started in manage.php, since
that's done in the imported header.php.  manage.php doesn't use that, so it
wasn't aware of user login state.

Fixed now.

Signed-off-by: Brian Warner <brian@bdwarner.com>