Ported setup file to Python, removed calendar table entirely, boosted speed
This is one of those backwards-compatibility-breaking kind of commits.  I
apologize in advance, but it comes with some sweet sweet goodness.

1) The setup file is now in Python. This was for a few reasons, one of the main
ones being consistency and readability. Goodbye, command line PHP.

2) The setup file can now do updates too, which leave the core project, repo,
settings, and tag data alone. It'll still dump all your data and recalculate it
next time, but this is actually the safest way to avoid stale data. It'll all
get rebuilt the next time you run facade-worker.py anyhow.

3) In the process, the calendar table stored procedure went away, and good
riddance to it. Now that the left join method of determining which repos are
missing analysis is gone, there was no reason to keep it. Plus, it's 10x faster
to generate the range of dates dynamically in Python.

4) The setup process is now (marginally) more secure, because it no longer echos
the administrator password back to you during input. Aaaaawkwaaaaard...

5) Just got rid of the setup directory. This stuff can live in utilities.

6) pip and a few other utilites are needed, so they are now in insteall-deps.sh

Signed-off-by: Brian Warner <brian@bdwarner.com>