Major change: results are now cached
As it turns out, in a big database, making many left joins is pretty expensive.
While the seems obvious - duh - I wasn't sure just how expensive they'd be.

Turns out, very.

This is a major revamp of the way data is retrieved by the web interface.  Now
instead of building the summary results tables dynamically (the expensive bit),
the results are cached after each facade-worker.py run.  This is possible and
better in retrospect because the results can't change unless facade-worker.py is
run anyhow.

In the process I also reworked the way that results are fetched.  Now it should
be possible to more easily sort results based upon something other than the
Total column.  This support will come later.

In addition, this lays the groundwork for a non-interactive kiosk display mode
that just runs off the cache files.  This would be useful if, for example, you
want to crunch data on a local machine where computation is free and
periodically push results to a remote server.

IMPORTANT NOTE: IF YOU ARE UPGRADING, RUN utilities/setup.py FIRST.  Otherwise
you will not have the cache tables in your database!

Signed-off-by: Brian Warner <brian@bdwarner.com>