The great big change that removes gitdm as a dependency
**** NOTE: THIS COMMIT IS NOT BACKWARDS COMPATIBLE. YOU NEED TO RECREATE YOUR
DATABASE USING utilities/setup.py AND SELECTING 'i' ****

(I'm sorry about that, but it's simply unavoidable. For the better, I promise.)

Facade was originally a front-end for gitdm, which provided the inspiration for
an analysis package like this. Over time I realized that with more data, more
insights could be drawn. As such, Facade now collects contributor data directly
from git log into the database.  This provides much more flexibility and enables
some additional features that weren't possible before.

On one hand, the setup is now much simpler, as we don't have to setup and
configure gitdm separately.  This also allows us to collect new statistics at
runtime that weren't tracked in gitdm, such as attempting to identify when an
addition/removal is simply modifying whitespace. We can now also split stats on
an author, committer, commit, or file basis. In the future, this paves the way
for file-level contributor analysis.

The way the analysis works has changed as well.  Previously, gitdm would crunch
through the git log for each day in the analysis period, and fill in any patches
committed that day. Now on each run we calculate all the parents of HEAD and
work backwards from there, filling in any commits which are missing. Admittedly
gitdm is still much faster for dense repos (like the kernel) but for slower
moving projects, this is much faster.

There are still a few things to be done, such as adding an interface to
add/remove domain and email mappings. It's possible to import config files in
the meantime, but this should be possible through the web interface, too. In
addition, I plan to add support for file-level analysis in the web UI.

I am grateful to Jon Corbet and the other authors of gitdm, as they were the
inspiration for this next evolution in Facade. It was this work that opened my
eyes as to the analysis possibilities, and what else could be done with the
data.

Signed-off-by: Brian Warner <brian@bdwarner.com>