Added add/remove/import/export aliases and affiliations
It's now possible to add and remove aliases and affiliations through the web UI.
In addition, you can download a CSV of the aliases and affiliations, and upload
a new version.  Existing aliases and affiliations are ignored in the import, so
if you want to change an alias you'll need to do it in the UI.

In addition, this triggered a change to the CSS which had to be rolled through
on the results page, too.

Signed-off-by: Brian Warner <brian@bdwarner.com>