Fixed emailmap import
There were some errors in the way the importer handled unexpected scenarios in
gitdm emailmap config files, such as bad behavior when entries were duplicated
or when no date range was given for overlapping domains/emails.

Also, the last entry to be imported was getting mangled.

Now the utility fixes the errors it can, and throws its hands up and exits for
those it can't. Should work better now.

Signed-off-by: Brian Warner <brian@bdwarner.com>