Fixes to alias discovery and management
Aliases weren't being handled in a very sustainable way.  Aliases were going
into the database as the canonical author or committer email address. While this
is great if the aliases are always right and the list is complete, it does us no
good if we have to change one because the original author/committer email is
gone.

Aliases are now discovered only for the purposes of assigning affiliations. This
keeps the original info in the database pure in case an alias needs to be
changed or removed.

Also, when adding or removing aliases and affiliations, the affected author or
committer affiliations are set to NULL so they are rediscovered next time
facade-worker.py runs.

Signed-off-by: Brian Warner <brian@bdwarner.com>