Automatically generate Excel summaries when facade-worker.py runs
facade-worker.py will now look for modules in excel_generators/ when it runs,
and execute any files in that directory that match generate*.py. This means that
if you don't like my default report, you can copy the template, fiddle with the
SQL statements, and save it as generate-super_awesome.py and the Excel file will
be automatically created.

Signed-off-by: Brian Warner <brian@bdwarner.com>