utf8 support in DB and automatic database, user, config file creation
I just learned that mysql doesn't use utf8 in default databases, but rather the
latin1 subset. This is a problem when using xlsxwriter, which really doesn't
like mangled characters.

The solution was the update the connectors in Python and PHP to use utf8mb4, and
to set utf8mb4 as the default character encoding on the database. Asking users
to do this is kind of a pain, so it's now possible to create the proper
database directly in the setup script, so long as the user has the root mysql
password.

A few other useful things were added in the process. setup.py can create the
database user and grant privileges. Username, password, and database name can
either be set by the user, or a random alphanumeric string. This means we're no
longer storing default connection info in a file somewhere.

In addition, setup.py writes both config files automatically, eliminating the
need to edit files by hand. From this commit forward, it should be sufficient to
run setup.py and be fully up and running.

Signed-off-by: Brian Warner <brian@bdwarner.com>