Add support for a different affiliations and aliases database
At times it can be helpful to spin up a secondary Facade instance, for example
if you want to quickly check stats but not go through a full update. You
probably want access to your affiliation and alias data, which you worked so
hard to build.  Now that's possible.

This commit touches a lot of files, but it adds a lot of functionality too:
 * setup.py now creates database credential files with the secondary 'people' db
 * All the php frontend supports a secondary 'people' db
 * Setup is a little more interactive when creating databases and users

Signed-off-by: Brian Warner <brian@bdwarner.com>