Add warnings to web UI when facade-worker.py is running
When facade-worker.py is running, the database is in a state of flux and any
data will be transitional. It could also have missing affiliations, if
facade-worker hasn't hit that part of the script yet.

A warning is now shown on the web UI that reports may be out of data, and
csv export is disabled unless the script is idle.

Signed-off-by: Brian Warner <brian@bdwarner.com>