Overhaul to affiliation and alias matching, and utf-8 fixes
Major overhaul to the way that aliases and affiliations are matched, added, and
removed. In addition, this commit adds true support for multiple facade
instances to use a single database for affiliation and alias data.

Previously when an alias or affiliation was changed, only the instance that made
the change would be aware of it (i.e., at the time of the change the contributor
info was set to NULL).  This would be detected by that instance the next time
facade-worker.py runs and filled, but any other database using that affiliation
and alias data would be unaware it needed to do something.

This change introduces a new column in the affiliation and alias tables called
'active', as well as a last modified timestamp.  In addition, each instance now
keeps track of the last time it checked for changes to the affiliation and alias
data.  When facade-worker.py runs, it checks to see what affiliation and alias
data has been added (set active to TRUE) or removed (set active to FALSE) and it
processes those changes at the time.  This allows each instance to process any
number of changes, and not necessarily in the same order.

This also includes some fixes for utf-8 support when importing and exporting
settings.  The more I dig into charset support, the more I desperately wish I
was still blissfully unaware of it. But it seems to be working for now.

Signed-off-by: Brian Warner <brian@bdwarner.com>