Add utf-8 BOMs to projects and settings csv file exports
I swear utf-8 will be the death of me. The sanity checks which ensure an
imported config file is what it purports to be were failing because Facade was
expecting a 3 byte offset and wasn't finding it. This is because I forgot to
write the BOM on export. This should work as expected, now.

Signed-off-by: Brian Warner <brian@bdwarner.com>