Suppress messages if table doesn't exist when using DROP TABLE IF EXISTS
This isn't strictly necessary, but it's super annoying when running the setup.py
script.

All tables are dropped during creation to clear out any stale data (the
user is warned appropriately).  However, if the table doesn't exist (which is
the case 99% of the time) mysql complains it couldn't delete the table. This
will cause user annoyance, so this patch nukes those warnings.

Signed-off-by: Brian Warner <brian@bdwarner.com>