Simplify check whether analysis has been run when showing results.
When there was a lot of data in the analysis table, the database query testing
whether analysis has been done took a long time to return.  This meant
intermittent 500 errors on the results page.

Since all we're actually doing is checking for the presence of analysis data in
the table, we can greatly shorten the query by only looking for one line of
data.

Signed-off-by: Brian Warner <brian@bdwarner.com>