Add support for Google Analytics tracking codes.
Google Analytics accounts can now be added on the configuration page.  The user
adds their code and the gtag.js code will be populated.

This required swapping the order of the header.php and db.php include, since we
now need to check whether a Google Analytics code was stored.

Signed-off-by: Brian Warner <brian@bdwarner.com>