Fix GitHub imports again
For sophisticated and complex technical reasons, GitHub imports had to be
disabled, due... um, to some... inconsis...

Ok fine, I disabled them because I didn't bother implementing any sort of
pagniation on calls to the GitHub API, and it was only returning the first 30
results, and I was lazy. Anyhow, GitHub import is back now and seems to work
well. Just be aware that GitHub has a rate limit for API calls, and if you try
to import a bunch of times it'll freeze you for a while.

Also, added the ability to bulk select all repos on the import page.

Signed-off-by: Brian Warner <brian@bdwarner.com>