Enable individual repositories to be updated when older than a set time
This patch enables the user to set a minimum wait time between when repos are
pulled and analyzed.  Repos that have been pulled within that amount of time are
skipped, and then checked again the next time facade-worker.py runs.

One major advantage of this is that facade-worker.py can now be run at much
shorter intervals without thrashing remote git repos.  When adding a new repo,
you don't have to wait as long for it to populate.

Future plans include adding the ability to manually flag a repo for updating in
the web UI.

Signed-off-by: Brian Warner <brian@bdwarner.com>