A better way to do database updates
After mulling it over, this backs out part of
82fb617c46205348431ed02d8a4fcf1dd3fb29d5 and moves the update functionality into
facade-worker itself.  Now each time Facade runs, it checks to see if the
database needs updating. This happens transparently, which is way better because
the user never needs to run setup.py to update the database. In theory at least,
facade-worker.py should never fail due to a database change - assuming the
update function works the way it should.

Bonus points, it detects databases the predate the versioning system, and
updates them too.

Signed-off-by: Brian Warner <brian@bdwarner.com>