When one repo updates, mark the whole project for an update
When we recache, we have to recache both the repo that changed and the entire
project. In a project with a lot of repos, this will mean a lot of duplicate
caching.  By aligning the repos to one facade-worker.py run, we do it all at
once and only cache the whole project once per waiting period.

Signed-off-by: Brian Warner <brian@bdwarner.com>