Improvements to worker that reduce duplicate operations
This patch improves the workflow considerably, and reduces certain types of
duplicate operations. Clone and pull operations were reversed, putting the pull
operation first so that updates in the projects' other repos are triggered by
adding a new one.  Also, once a repo is cloned it's immediately marked for
analysis, while the other repos in the project are marked to update.

Signed-off-by: Brian Warner <brian@bdwarner.com>