Database changes to support future multithreading
NOTE: This commit alters the database, so it will trigger a new major version.

The only thing holding us back from doing multithreading was the way
we stored information on which commit was currently being processed by
facade-worker.py.  Previously there was a single column in repos where
this info was stored.  This was great when only one commit was being
analyzed at a time, but wouldn't work at all if commits were going to be
analyzed in parallel.  If the program exited abnormally and multiple
threads were active, only the last one to store its working commit could
be detected and cleaned up.

The solution was to move the working commits into their own table, so
now we can have an arbitrary number of working commits per repo.

Signed-off-by: Brian Warner <brian@bdwarner.com>