Add flag to manually trigger an analysis
To save time, if a git repo hasn't updated, Facade won't analyze it for
missing commits.  Under normal circumstances this is ok (for example, if
an analysis is interrupted, the repo will not be marked as complete).
However, if you're debugging, it can be nice to force Facade to analyze
a repo for missing commits even if it hasn't been updated.

Signed-off-by: Brian Warner <brian@bdwarner.com>