Import databases in a more sane way
Previously we imported our database and cursor from a module which was
created by setup.py. This worked, but only if you were happy with just
those database and cursors and never wanted new ones.

Turns out, MySQLdb is not threadsafe, so if we want to do
multithreading we'll need to open a new database connection for each
patch that we analyze.  Now instead of importing the database and the
cursor, we just import the credentials directly, so they're easier to
use.

When facade-worker.py is run the first time, it will detect the old
db.py, import the credentials, create the new db.cfg, and delete the old
files.  This is all transparent to the user.

This commit will trigger a new major version when merged to master,
because it changes the way that a database is accessed.

Signed-off-by: Brian Warner <brian@bdwarner.com>