The multithreading update
Up to this point, the analysis has basically been a bunch of independent
operations in a straight line. As it turns out, it lends itself very
well to multithreading, as each commit in a repo is independent of every
other commit, so long as you keep track of which ones you need to
scrape.

There are quite a few changes in this patch, but they are all related.
In particular:
 * Add multithreading support for analysis
 * Add the ability to disable multithreading if you want to
 * Big reorganization of functions, because MySQLdb is not threadsafe

On the last point in particular, many global convenience functions are
now local.  This was essential for anything that required database
access, so that it would use the local database connection and cursor
instead of deadlocking on the global one.

This turned into a general reorganization, as it made more sense to keep
convenience functions closer to where they were being called from.  Over
time as the program got longer and longer they had drifted apart.

This also lays some groundwork should we ever decide to try
multithreading on other functions.

Signed-off-by: Brian Warner <brian@bdwarner.com>