Add PyPy3 support and smarter db connections
This patch should have performance implications. I have wanted to add
PyPy support for a while, because it generally has pretty significant
performance consequences. The blockers have been getting the module
dependencies resolved in an easy and consistent way for users (see
configure_pypy3.sh) and dealing with the database client issue.

MySQLdb is a nice fast database client, but it is incompatible with
PyPy. pymysql is a nice native Python database client, but it's really
slow when compared with MySQLdb.

With this patch, facade-worker.py checks to see which interpreter called
it, and chooses the database client accordingly.

In the process of making this change, I also added some logic that
decides whether we need a new local database connection during the
analysis of each commit, based upon whether we are multithreading. This
is a much smarter way to handle database connectionsi. When Facade is
being run in singlethreaded mode threadsafety isn't a concern, so
there's no reason to be penalized by opening unnecessary connections
when the general purpose connection will do.

Signed-off-by: Brian Warner <brian@bdwarner.com>