Reorganize web files into a subdirectory (!!!!)
>> THIS IS A CHANGE THAT REQUIRES MANUAL INTERVENTION <<

If you have an existing install, you will need to update the following
files:

/etc/apache2/apache2.conf
 <Directory /path/to/facade/> --> <Directory /path/to/facade/web/>

/etc/apache2/sites-available/facade.conf
 DocumentRoot /path/to/facade/ --> DocumentRoot /path/to/facade/web/

In both cases, you are just adding 'web/' to the end of the path to
reflect the new home of the web UI files.

Now, onto the reason for this.  The web UI has always been separate and
discrete from the Python files that do the actual work.  Aside from
feeding configuration data in and displaying the results, the web stuff
is not a critical part of Facade.

By consolidating the files in a separate subdirectory, it makes this
more apparent and should make it easier to add new UIs.

Signed-off-by: Brian Warner <brian@bdwarner.com>