Simplify getting database credentials in PHP
Previously we had created two separate database files, one for PHP and
one for Python.  With the change to using an ini-style file, this became
a source of error.  Most notably, in the recent release that
consolidated the PHP files in a different directory, which required the
user to move the previously generated creds.php file into the right
place. Of course, I forgot to put this in the commit log...

This change is a bugfix which just tells PHP to look for the credentials
in a different place.  The existing creds.php file is now unused.  A
future patch will remove the process of writing creds.php from setup.py.
For now, this will just make the update process more transparent.

Signed-off-by: Brian Warner <brian@bdwarner.com>