Improve settings export and import
When a user changes a Facade setting, the old settings are kept in the
settings table for auditability. When Facade needs a setting, it grabs
the one with the most recent timestamp.  However, when importing on a
very fast server, it's possible the settings could be added so quickly
that the timestamps are indistinguishable.

To deal with this, we're now exporting settings in order of their ID,
which ensures they're read in order as well.  When importing, there's
now a very short delay between lines, ensuring that there's a meaningful
difference between the timestamps.

Signed-off-by: Brian Warner <brian@bdwarner.com>