Fix unique constraints on tags
Unbeknownst to me, when you have a NULL value in a MySQL constraint,
you're allowed to have multiple matches.  What this means is that if you
have a UNIQUE constraint on a table where three columns are mandatory
and the fourth is not, it'll keep merrily adding duplicate rows where
the mandatory ones are the same and the fourth is empty.  This isn't the
behavior we wanted, as it allows duplicate tags.

The solution was the remove the old constraint and add a new one that
excludes the column that allows NULL values.  Yes, it's arguably less
strict, but it has the major advantage of actually working, so it has
that going for it.

Signed-off-by: Brian Warner <brian@bdwarner.com>