Initial commit of the CLI / importable module
This is the initial commit of the Facade CLI.  It's a 100% Python3 user
interface which can be used instead of the PHP-based web interface. It
is mostly at feature parity with the web UI. The main things missing
are:
 * Project-specific stats (you'll have to export an Excel file)
 * Bulk repo imports from GitHub, cgit, and Gerrit
 * Filtered exports

Otherwise, the other major features are there.

In addition, it's possible to import cli/facade.py using other Python
programs, and it'll provide some useful basic management functions.
These are documented at the head of the file.

Future tasks include:
 * Adding the ability to do reports
 * Adding the ability to do bulk repo imports
 * Adding export filters
 * Adding a setup option for CLI-only that doesn't require Apache

This will be a work in progress, but the goal is to maintain it in
parallel to the web interface.

Signed-off-by: Brian Warner <brian@bdwarner.com>