Optimize tables when deleting repos
When you delete large volumes of data from mysql (for example, you
removed a repo or project and analysis data is removed from
analysis_data and the caches) it doesn't reclaim the disk space.  For
most use cases this is probably not a big deal, but for long-lived
installations it could cause performance and bloating issues.

Now when a project or repo is deleted, we optimize the tables.  This
will increase the run time during deletions, but should improve
performance long term.

Signed-off-by: Brian Warner <brian@bdwarner.com>