Add a minimal install option (no Apache)
Now that we have a CLI, the web UI is optional. Well, optional in the
sense that you don't strictly need it, but truth be told it is still a
lot nicer to use.  But, maybe you don't want to run a web server on your
machine, or maybe you want to do the analysis on one system and display
it using another.  This should all be possible now.

Signed-off-by: Brian Warner <brian@bdwarner.com>