Modify handling of end dates in special_tags
When originally written, a special tag without an explicit end date was
given a NULL end_date in the database.  This meant that we had to check
for an explicit end_date and if it wasn't expected, had to do some
contortions to change the SQL queries around.

It is much more straightforward to choose an end date that is just
impossibly far into the future.  This makes math much easier when
matching tags to data points.

Signed-off-by: Brian Warner <brian@bdwarner.com>