Import more than 500 repos at a time
Turns out, PHP limits the number of items you can parse in a POST request.
There are good reasons they did this, but in this case we sometimes want to
import a massive number of repos.  This commit makes a tweak to the .htaccess
file allowing import of up to 2,499 repos at once (in theory).

Signed-off-by: Brian Warner <brian@bdwarner.com>